Move different part (DGUS, Logs, ...) into their own files.
